pyproject.toml
[project]
name = "reexample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-analytics-data>=0.20.0",
    "monsterui>=1.0.42",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.40",
    "supabase>=2.0.0",
]
